refact: restructure project
1. tests/src/qserializer/tests/types
   -> tests/src/qserializer/test_types;
2. use ${PROJECT_SOURCE_DIR};
3. use OBJECT library for header only library;
4. make every directory a build target.
