+++ CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required (
         VERSION
-                3.23
+                3.24
 )
 
 project (
