+++ tests/src/qserializer/test_types/Base.cpp
@@ -1,0 +1,5 @@
+#include "qserializer/test_types/Base.h"
+
+#include "qserializer/core.h" // for QSERIALIZER_IMPL
+
+QSERIALIZER_IMPL(qserializer::test_types::Base);
