+++ tests/src/qserializer/test_utils.cpp
@@ -1,0 +1,1 @@
+#include "qserializer/test_utils.h" // IWYU pragma: keep
