--- tests/src/qserializer/tests/types/Base.cpp
@@ -1,5 +1,0 @@
-#include "qserializer/tests/types/Base.h"
-
-#include "qserializer/core.h" // for QSERIALIZER_IMPL
-
-QSERIALIZER_IMPL(qserializer::tests::types::Base);
