--- tests/src/qserializer/tests/types/Page.cpp
@@ -1,5 +1,0 @@
-#include "qserializer/tests/types/Page.h"
-
-#include "qserializer/core.h" // for QSERIALIZER_IMPL
-
-QSERIALIZER_IMPL(qserializer::tests::types::Page);
