--- tests/src/qserializer/tests/utils.check.cpp
@@ -1,1 +1,0 @@
-#include "qserializer/tests/utils.h" // IWYU pragma: keep
